# apps/api/src/tickets/tickets.controller.ts
import { ApiResource, ApiResult } from '../openapi/api-documentation.js';
import {
  Body,
  Controller,
  ForbiddenException,
  Get,
  HttpCode,
  Param,
  ParseUUIDPipe,
  Patch,
  Post,
  Req,
  UseGuards,
  Delete,
  Query,
} from '@nestjs/common';
import { minutes, Throttle } from '@nestjs/throttler';

import { PermissionGuard } from '../rbac/permission.guard.js';
import { PERMISSIONS } from '../rbac/permissions.js';
import { RequirePermissions } from '../rbac/require-permissions.decorator.js';
import { TenantMembershipGuard } from '../tenancy/tenant-membership.guard.js';
import type { TenantAuthenticatedRequest } from '../tenancy/tenant-context.types.js';
import { CreateTicketDto } from './dto/create-ticket.dto.js';
import { TicketsService } from './tickets.service.js';
import { UpdateTicketDto } from './dto/update-ticket.dto.js';
import { UpdateTicketStatusDto } from './dto/update-ticket-status.dto.js';
import { UpdateTicketAssigneeDto } from './dto/update-ticket-assignee.dto.js';
import { CreateTicketMessageDto } from './dto/create-ticket-message.dto.js';
import {ListTicketsDto} from './dto/list-tickets.dto.js';

@ApiResource('Tickets')
@Controller(
  'v1/organizations/:organizationId/tickets',
)
@UseGuards(
  TenantMembershipGuard,
  PermissionGuard,
)
export class TicketsController {
  constructor(
    private readonly ticketsService:
      TicketsService,
  ) {}

 @ApiResult('Tickets')
  @Get()
@RequirePermissions(
  PERMISSIONS.TICKETS_READ,
)
async list(
  @Req()
  request:
    TenantAuthenticatedRequest,

  @Query()
  query:
    ListTicketsDto,
) {
  const tenant =
    request.tenant;

  if (!tenant) {
    throw new ForbiddenException(
      'Tenant context is required',
    );
  }

  return this.ticketsService.list(
    tenant.organizationId,
    query,
  );
}

  @ApiResult('Ticket')
  @Post()
  @RequirePermissions(
    PERMISSIONS.TICKETS_WRITE,
  )
  async create(
    @Req()
    request:
      TenantAuthenticatedRequest,

    @Body()
    dto: CreateTicketDto,
  ) {
    const tenant =
      request.tenant;

    if (!tenant) {
      throw new ForbiddenException(
        'Tenant context is required',
      );
    }

    return this.ticketsService.create(
      tenant,
      dto,
    );
  }

    @ApiResult('Ticket')
  @Get(':ticketId')
    @RequirePermissions(
    PERMISSIONS.TICKETS_READ,
    )
    async getOne(
    @Req()
    request:
        TenantAuthenticatedRequest,

    @Param(
        'ticketId',
        new ParseUUIDPipe(),
    )
    ticketId: string,
    ) {
    const tenant =
        request.tenant;

    if (!tenant) {
        throw new ForbiddenException(
        'Tenant context is required',
        );
    }

    return this.ticketsService.findOne(
        tenant.organizationId,
        ticketId,
    );
}

@ApiResult('Ticket')
  @Patch(':ticketId')
@RequirePermissions(
  PERMISSIONS.TICKETS_WRITE,
)
async update(
  @Req()
  request:
    TenantAuthenticatedRequest,

  @Param(
    'ticketId',
    new ParseUUIDPipe(),
  )
  ticketId: string,

  @Body()
  dto: UpdateTicketDto,
) {
  const tenant =
    request.tenant;

  if (!tenant) {
    throw new ForbiddenException(
      'Tenant context is required',
    );
  }

  return this.ticketsService.update(
    tenant,
    ticketId,
    dto,
  );
}

@ApiResult('Ticket')
  @Patch(':ticketId/status')
@RequirePermissions(
  PERMISSIONS.TICKETS_WRITE,
)
async updateStatus(
  @Req()
  request:
    TenantAuthenticatedRequest,

  @Param(
    'ticketId',
    new ParseUUIDPipe(),
  )
  ticketId: string,

  @Body()
  dto:
    UpdateTicketStatusDto,
) {
  const tenant =
    request.tenant;

  if (!tenant) {
    throw new ForbiddenException(
      'Tenant context is required',
    );
  }

  return this.ticketsService
    .updateStatus(
      tenant,
      ticketId,
      dto.status,
    );
}

@ApiResult('Ticket')
  @Patch(':ticketId/assignee')
@RequirePermissions(
  PERMISSIONS.TICKETS_WRITE,
)
async assign(
  @Req()
  request:
    TenantAuthenticatedRequest,

  @Param(
    'ticketId',
    new ParseUUIDPipe(),
  )
  ticketId: string,

  @Body()
  dto:
    UpdateTicketAssigneeDto,
) {
  const tenant =
    request.tenant;

  if (!tenant) {
    throw new ForbiddenException(
      'Tenant context is required',
    );
  }

  return this.ticketsService.assign(
    tenant,
    ticketId,
    dto.membershipId,
  );
}

@ApiResult('Ticket')
  @Post(':ticketId/tags/:tagId')
@HttpCode(200)
@RequirePermissions(
  PERMISSIONS.TICKETS_WRITE,
)
async addTag(
  @Req()
  request:
    TenantAuthenticatedRequest,

  @Param(
    'ticketId',
    new ParseUUIDPipe(),
  )
  ticketId: string,

  @Param(
    'tagId',
    new ParseUUIDPipe(),
  )
  tagId: string,
) {
  const tenant =
    request.tenant;

  if (!tenant) {
    throw new ForbiddenException(
      'Tenant context is required',
    );
  }

  return this.ticketsService.addTag(
    tenant,
    ticketId,
    tagId,
  );
}

@ApiResult('Ticket')
  @Delete(':ticketId/tags/:tagId')
@RequirePermissions(
  PERMISSIONS.TICKETS_WRITE,
)
async removeTag(
  @Req()
  request:
    TenantAuthenticatedRequest,

  @Param(
    'ticketId',
    new ParseUUIDPipe(),
  )
  ticketId: string,

  @Param(
    'tagId',
    new ParseUUIDPipe(),
  )
  tagId: string,
) {
  const tenant =
    request.tenant;

  if (!tenant) {
    throw new ForbiddenException(
      'Tenant context is required',
    );
  }

  return this.ticketsService.removeTag(
    tenant,
    ticketId,
    tagId,
  );
}

@ApiResult('Messages')
  @Get(':ticketId/messages')
@RequirePermissions(
  PERMISSIONS.TICKETS_READ,
)
async listMessages(
  @Req()
  request:
    TenantAuthenticatedRequest,

  @Param(
    'ticketId',
    new ParseUUIDPipe(),
  )
  ticketId: string,
) {
  const tenant =
    request.tenant;

  if (!tenant) {
    throw new ForbiddenException(
      'Tenant context is required',
    );
  }

  return this.ticketsService.listMessages(
    tenant.organizationId,
    ticketId,
  );
}

@ApiResult('Message')
  @Post(':ticketId/messages')
@Throttle({
  default: {
    limit: 30,
    ttl: minutes(1),
  },
})
@RequirePermissions(
  PERMISSIONS.TICKETS_WRITE,
)
async createMessage(
  @Req()
  request:
    TenantAuthenticatedRequest,

  @Param(
    'ticketId',
    new ParseUUIDPipe(),
  )
  ticketId: string,

  @Body()
  dto:
    CreateTicketMessageDto,
) {
  const tenant =
    request.tenant;

  if (!tenant) {
    throw new ForbiddenException(
      'Tenant context is required',
    );
  }

  return this.ticketsService.createMessage(
    tenant,
    ticketId,
    dto,
  );
}
}