# docs/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/organizations/{organizationId}/tickets": {
      "get": {
        "description": "Requires a verified Auth0 access token for the configured API audience. Organization membership is resolved server-side. A missing membership or foreign resource returns 404; IDs never grant access. Required permissions: tickets:read. See docs/security.md for the role matrix and service-level restrictions. DTO validation and documented schemas do not replace authentication or business rules.",
        "operationId": "TicketsController_list",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "integer"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "integer"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "maxLength": 200,
              "nullable": true,
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "nullable": true,
              "type": "string",
              "enum": [
                "OPEN",
                "PENDING",
                "RESOLVED",
                "CLOSED"
              ]
            }
          },
          {
            "name": "priority",
            "required": false,
            "in": "query",
            "schema": {
              "nullable": true,
              "type": "string",
              "enum": [
                "LOW",
                "NORMAL",
                "HIGH",
                "URGENT"
              ]
            }
          },
          {
            "name": "customerId",
            "required": false,
            "in": "query",
            "schema": {
              "format": "uuid",
              "nullable": true,
              "type": "string"
            }
          },
          {
            "name": "assigneeMembershipId",
            "required": false,
            "in": "query",
            "schema": {
              "format": "uuid",
              "nullable": true,
              "type": "string"
            }
          },
          {
            "name": "tagId",
            "required": false,
            "in": "query",
            "schema": {
              "format": "uuid",
              "nullable": true,
              "type": "string"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "schema": {
              "default": "updatedAt",
              "type": "string",
              "enum": [
                "createdAt",
                "updatedAt"
              ]
            }
          },
          {
            "name": "sortOrder",
            "required": false,
            "in": "query",
            "schema": {
              "default": "desc",
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            }
          },
          {
            "name": "organizationId",
            "required": true,
            "in": "path",
            "description": "Tenant-scoped UUID. Ownership and membership are checked server-side.",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Tickets"
                }
              }
            }
          },
          "400": {
            "description": "Malformed identifiers, DTO validation failure or invalid workflow.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid, expired or wrong-audience access token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated member lacks permission or violates a role restriction.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Organization membership or tenant-scoped resource not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "auth0": []
          }
        ],
        "summary": "Tickets: list",
        "tags": [
          "Tickets"
        ]
      },
      "post": {
        "description": "Requires a verified Auth0 access token for the configured API audience. Organization membership is resolved server-side. A missing membership or foreign resource returns 404; IDs never grant access. Required permissions: tickets:write. See docs/security.md for the role matrix and service-level restrictions. DTO validation and documented schemas do not replace authentication or business rules.",
        "operationId": "TicketsController_create",
        "parameters": [
          {
            "name": "organizationId",
            "required": true,
            "in": "path",
            "description": "Tenant-scoped UUID. Ownership and membership are checked server-side.",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTicketDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Ticket"
                }
              }
            }
          },
          "400": {
            "description": "Malformed identifiers, DTO validation failure or invalid workflow.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid, expired or wrong-audience access token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated member lacks permission or violates a role restriction.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Organization membership or tenant-scoped resource not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Conflict with existing data or workflow state where applicable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "auth0": []
          }
        ],
        "summary": "Tickets: create",
        "tags": [
          "Tickets"
        ]
      }
    },
    "/v1/organizations/{organizationId}/tickets/{ticketId}": {
      "get": {
        "description": "Requires a verified Auth0 access token for the configured API audience. Organization membership is resolved server-side. A missing membership or foreign resource returns 404; IDs never grant access. Required permissions: tickets:read. See docs/security.md for the role matrix and service-level restrictions. DTO validation and documented schemas do not replace authentication or business rules.",
        "operationId": "TicketsController_getOne",
        "parameters": [
          {
            "name": "ticketId",
            "required": true,
            "in": "path",
            "description": "Tenant-scoped UUID. Ownership and membership are checked server-side.",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "organizationId",
            "required": true,
            "in": "path",
            "description": "Tenant-scoped UUID. Ownership and membership are checked server-side.",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Ticket"
                }
              }
            }
          },
          "400": {
            "description": "Malformed identifiers, DTO validation failure or invalid workflow.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid, expired or wrong-audience access token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated member lacks permission or violates a role restriction.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Organization membership or tenant-scoped resource not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "auth0": []
          }
        ],
        "summary": "Tickets: get one",
        "tags": [
          "Tickets"
        ]
      },
      "patch": {
        "description": "Requires a verified Auth0 access token for the configured API audience. Organization membership is resolved server-side. A missing membership or foreign resource returns 404; IDs never grant access. Required permissions: tickets:write. See docs/security.md for the role matrix and service-level restrictions. DTO validation and documented schemas do not replace authentication or business rules.",
        "operationId": "TicketsController_update",
        "parameters": [
          {
            "name": "ticketId",
            "required": true,
            "in": "path",
            "description": "Tenant-scoped UUID. Ownership and membership are checked server-side.",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "organizationId",
            "required": true,
            "in": "path",
            "description": "Tenant-scoped UUID. Ownership and membership are checked server-side.",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTicketDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Ticket"
                }
              }
            }
          },
          "400": {
            "description": "Malformed identifiers, DTO validation failure or invalid workflow.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid, expired or wrong-audience access token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated member lacks permission or violates a role restriction.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Organization membership or tenant-scoped resource not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Conflict with existing data or workflow state where applicable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "auth0": []
          }
        ],
        "summary": "Tickets: update",
        "tags": [
          "Tickets"
        ]
      }
    },
    "/v1/organizations/{organizationId}/tickets/{ticketId}/status": {
      "patch": {
        "description": "Requires a verified Auth0 access token for the configured API audience. Organization membership is resolved server-side. A missing membership or foreign resource returns 404; IDs never grant access. Required permissions: tickets:write. See docs/security.md for the role matrix and service-level restrictions. DTO validation and documented schemas do not replace authentication or business rules.",
        "operationId": "TicketsController_updateStatus",
        "parameters": [
          {
            "name": "ticketId",
            "required": true,
            "in": "path",
            "description": "Tenant-scoped UUID. Ownership and membership are checked server-side.",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "organizationId",
            "required": true,
            "in": "path",
            "description": "Tenant-scoped UUID. Ownership and membership are checked server-side.",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTicketStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Ticket"
                }
              }
            }
          },
          "400": {
            "description": "Malformed identifiers, DTO validation failure or invalid workflow.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid, expired or wrong-audience access token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated member lacks permission or violates a role restriction.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Organization membership or tenant-scoped resource not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Conflict with existing data or workflow state where applicable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "auth0": []
          }
        ],
        "summary": "Tickets: update status",
        "tags": [
          "Tickets"
        ]
      }
    },
    "/v1/organizations/{organizationId}/tickets/{ticketId}/assignee": {
      "patch": {
        "description": "Requires a verified Auth0 access token for the configured API audience. Organization membership is resolved server-side. A missing membership or foreign resource returns 404; IDs never grant access. Required permissions: tickets:write. See docs/security.md for the role matrix and service-level restrictions. DTO validation and documented schemas do not replace authentication or business rules.",
        "operationId": "TicketsController_assign",
        "parameters": [
          {
            "name": "ticketId",
            "required": true,
            "in": "path",
            "description": "Tenant-scoped UUID. Ownership and membership are checked server-side.",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "organizationId",
            "required": true,
            "in": "path",
            "description": "Tenant-scoped UUID. Ownership and membership are checked server-side.",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTicketAssigneeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Ticket"
                }
              }
            }
          },
          "400": {
            "description": "Malformed identifiers, DTO validation failure or invalid workflow.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid, expired or wrong-audience access token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated member lacks permission or violates a role restriction.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Organization membership or tenant-scoped resource not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Conflict with existing data or workflow state where applicable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "auth0": []
          }
        ],
        "summary": "Tickets: assign",
        "tags": [
          "Tickets"
        ]
      }
    },
...
  },
  "info": {
    "title": "OpsDesk API",
    "description": "Tenant-scoped support operations API. Use the Next.js BFF for browser requests; Auth0 bearer tokens belong on the API boundary. See docs/api/README.md for workflows and Socket.IO events. No real credentials or account details are embedded.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://127.0.0.1:3001",
      "description": "Local API only; configure deployment URLs privately"
    }
  ],
  "components": {
    "securitySchemes": {
      "auth0": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Auth0 access token with the configured API audience and issuer; never an ID token."
      },
      "svix": {
        "type": "apiKey",
        "in": "header",
        "name": "svix-signature",
        "description": "Verified Resend/Svix signature. Also requires svix-id and svix-timestamp; the signature covers the exact raw body."
      }
    },
    "schemas": {
...
      "CreateTicketDto": {
        "type": "object",
        "properties": {
          "customerId": {
            "type": "string",
            "format": "uuid"
          },
          "subject": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200
          },
          "description": {
            "type": "string",
            "nullable": true,
            "maxLength": 5000
          },
          "priority": {
            "type": "string",
            "nullable": true,
            "enum": [
              "LOW",
              "NORMAL",
              "HIGH",
              "URGENT"
            ]
          }
        },
        "required": [
          "customerId",
          "subject"
        ],
        "additionalProperties": false
      },
      "UpdateTicketDto": {
        "type": "object",
        "properties": {
          "subject": {
            "type": "string",
            "nullable": true,
            "minLength": 1,
            "maxLength": 200
          },
          "description": {
            "type": "string",
            "nullable": true,
            "maxLength": 5000
          },
          "priority": {
            "type": "string",
            "nullable": true,
            "enum": [
              "LOW",
              "NORMAL",
              "HIGH",
              "URGENT"
            ]
          }
        },
        "additionalProperties": false
      },
      "UpdateTicketStatusDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "OPEN",
              "PENDING",
              "RESOLVED",
              "CLOSED"
            ]
          }
        },
        "required": [
          "status"
        ],
        "additionalProperties": false
      },
      "UpdateTicketAssigneeDto": {
        "type": "object",
        "properties": {
          "membershipId": {
            "type": "string",
            "nullable": true,
            "format": "uuid",
            "description": "Assignable membership in this organization; null clears the assignee. VIEWER memberships cannot be assigned."
          }
        },
        "required": [
          "membershipId"
        ],
        "additionalProperties": false
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "integer"
          },
          "message": {
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            ]
          },
          "error": {
            "type": "string"
          }
        },
        "required": [
          "statusCode",
          "message"
        ]
      },
...
    }
  }
}